+++ ports/esp8266/nodemcu/main/sleep.c
@@ -2,7 +2,7 @@
 #include "freertos/FreeRTOS.h"
 #include "freertos/task.h"
 
-void sleep_ms(unsigned long msec)
+void sleep_ms(unsigned int msec)
 {
 	vTaskDelay(msec / portTICK_PERIOD_MS);
 }
